# Conflicts:
#	src/main/java/com/green/muziuniv_be_notuser/openfeign/user/UserClient.java

+++ src/main/java/com/green/muziuniv_be_notuser/app/shared/application/model/AppPostReq.java
@@ -9,11 +9,11 @@ import lombok.Setter;
 @NoArgsConstructor
 @AllArgsConstructor
 public class AppPostReq {
-    private int userId;
+    private Long userId;
     private int scheduleId;
     private String reason;
 
-    public void setUserId(int userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 }
